## Overview

This commit optimizes `SHOW TAG VALUES` so that it avoids the `SELECT` query engine execution and iterator creation. There are also optimizations to reduce individual memory allocations and to reduce in-memory heap size by only operating on one measurement at a time.

Execution time has been reduce to approximately 900ms for 500,000 rows. This is about 2µs per row. Of this time, approximately 1µs is spent retrieving and sorting the row and 1µs is spent encoding into JSON and writing to the response body.

Fixes #6233

/cc @pauldix @jwilder 


### Notes

The previous implementation of `tagValuesIterator` still exists because it is used for the distributed `SHOW TAG VALUES`. This change only affects the single node implementation.

## TODO

- [x] Rebased/mergable
- [x] Tests pass
- [x] Sign [CLA](https://influxdata.com/community/cla/) (if not already signed)