@jwilder 2016-06-06
I think those unexported funcs might need to stay unexported and unlocked and instead create a new exported version that takes a read-lock?